+++ src/renderer/components/helpers/ChatMethods.ts
@@ -1,4 +1,3 @@
-import { sendMessageParams } from '../../delta-remote'
 import ChatStore, { ChatView } from '../../stores/chat'
 import { ScreenContext, unwrapContext } from '../../contexts'
 import { C } from 'deltachat-node/node/dist/constants'
@@ -11,6 +10,7 @@ import ViewGroup from '../dialogs/ViewGroup'
 import ViewProfile from '../dialogs/ViewProfile'
 import ConfirmationDialog from '../dialogs/ConfirmationDialog'
 import { T } from '@deltachat/jsonrpc-client'
+import { sendMessageParams } from '../../../shared/shared-types'
 
 const log = getLogger('renderer/message')
