--- src/renderer/delta-remote.ts
@@ -1,21 +1,0 @@
-import { _callDcMethodAsync } from './ipc'
-
-export type sendMessageParams = {
-  text?: string | null
-  filename?: string
-  location?: {
-    lat: number
-    lng: number
-  }
-  quoteMessageId?: number
-}
-
-class DeltaRemote {
-  // catchall: ----------------------------------------------------------
-  call(fnName: string): Promise<any>
-  call(fnName: string, ...args: any[]): Promise<any> {
-    return _callDcMethodAsync(fnName, ...args)
-  }
-}
-
-export const DeltaBackend = new DeltaRemote()
