+++ src/renderer/stores/chat.ts
@@ -1,5 +1,4 @@
 import { Store, useStore } from './store'
-import { sendMessageParams } from '../delta-remote'
 import { ActionEmitter, KeybindAction } from '../keybindings'
 import { C } from 'deltachat-node/node/dist/constants'
 import { OrderedMap } from 'immutable'
@@ -16,6 +15,7 @@ import {
 import { ChatStoreScheduler } from './chat/chat_scheduler'
 import { saveLastChatId } from './chat/chat_sideeffects'
 import { onReady } from '../onready'
+import { sendMessageParams } from '../../shared/shared-types'
 
 export const PAGE_SIZE = 11
